Move CatalogCleaner to PsulibBlacklight::DataManager
This does a couple of things:

1. Fixes a bug when running rake tasks: these need to be loaded in
   either dev or test.

2. Removes the ci task: everything is now defined _explicitly_ in
   bin/ci-rspec

3. Re-names tasks to their proper namespace

4. Removes docker-related tasks: these aren't needed anymore since we're
   using docker-compose.

5. Removes Procfile and foreman gem: we're not using these anymore to
   manage our local dev environment.
