+++ Gemfile
@@ -35,7 +35,6 @@ gem 'blacklight_range_limit'
 group :development do
   gem 'better_errors'
   gem 'binding_of_caller'
-  gem 'foreman', '~> 0.63.0'
   gem 'listen'
   gem 'spring'
   gem 'spring-commands-rspec'
