--- Procfile.dev
@@ -1,4 +1,0 @@
-solr: bundle exec rails solr:up
-web: rails s -p 3000
-webpacker: bin/webpack-dev-server
-redis: redis-server
