+++ Rakefile
@@ -7,15 +7,6 @@ require_relative 'config/application'
 
 unless Rails.env.production?
   Rails.root.join('tasks').children.map { |file| load(file) }
-
-  desc 'Run continuous integration build'
-  task ci: :environment do
-    Rake::Task['blackcat:solr:index'].invoke
-    Rake::Task['spec'].invoke
-    Rake::Task['blackcat:solr:deindex'].invoke
-  end
-
-  task default: :ci
 end
 
 Rails.application.load_tasks
