+++ bin/ci-rspec
@@ -11,6 +11,7 @@ echo "initalizing solr"
 bundle exec rake solr:create_collection
 bundle exec rake solr:update_config
 bundle exec rake solr:create_alias
+bundle exec rake solr:load_fixtures
 
 
 echo "Downloading webdrivers"
@@ -27,7 +28,7 @@ if [ ! -f ${REPORTER_BIN} ]; then
 fi
 
 ./${REPORTER_BIN} before-build
-bundle exec rails ci
+bundle exec rspec
 RSPEC_EXIT_CODE=$?
 ./${REPORTER_BIN} after-build
 exit ${RSPEC_EXIT_CODE}
